chore: export reusable useCopy hook
